# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "crime-intelligence-platform",
    "description": "Crime Intelligence Platform API for Karnataka State Police. Frontend contract: use /docs or /openapi.json. AI routes use Catalyst QuickML only.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analytics/geo/incidents": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Geo Incidents",
        "description": "Lat/long incident pins for the map.",
        "operationId": "geo_incidents_api_v1_analytics_geo_incidents_get",
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          },
          {
            "name": "police_station_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Police Station Id"
            }
          },
          {
            "name": "case_status_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Status Id"
            }
          },
          {
            "name": "crime_major_head_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crime Major Head Id"
            }
          },
          {
            "name": "registered_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered From"
            }
          },
          {
            "name": "registered_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentPointsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/hotspots": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Hotspots",
        "description": "Spatiotemporal hotspot bins (location \u00d7 optional hour-of-day).",
        "operationId": "hotspots_api_v1_analytics_hotspots_get",
        "parameters": [
          {
            "name": "cell_size_degrees",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.01,
              "default": 0.05,
              "title": "Cell Size Degrees"
            }
          },
          {
            "name": "grain",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "hour",
                "day"
              ],
              "type": "string",
              "default": "hour",
              "title": "Grain"
            }
          },
          {
            "name": "registered_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered From"
            }
          },
          {
            "name": "registered_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Registered To"
            }
          },
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/alerts/trends": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Trend Alerts",
        "description": "Emerging trend alerts: recent volume vs baseline by district + crime head.",
        "operationId": "trend_alerts_api_v1_analytics_alerts_trends_get",
        "parameters": [
          {
            "name": "recent_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 7,
              "title": "Recent Days"
            }
          },
          {
            "name": "baseline_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 28,
              "title": "Baseline Days"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 10.0,
              "minimum": 1.0,
              "default": 1.5,
              "title": "Threshold"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendAlertsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HotspotBin": {
        "properties": {
          "lat_bin": {
            "type": "number",
            "title": "Lat Bin"
          },
          "lon_bin": {
            "type": "number",
            "title": "Lon Bin"
          },
          "hour_of_day": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hour Of Day",
            "description": "0-23 when grain includes hour; null if day-only"
          },
          "case_count": {
            "type": "integer",
            "title": "Case Count"
          },
          "sample_case_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Sample Case Ids"
          }
        },
        "type": "object",
        "required": [
          "lat_bin",
          "lon_bin",
          "case_count"
        ],
        "title": "HotspotBin",
        "description": "Spatiotemporal cluster cell."
      },
      "HotspotsResponse": {
        "properties": {
          "grain": {
            "type": "string",
            "title": "Grain"
          },
          "cell_size_degrees": {
            "type": "number",
            "title": "Cell Size Degrees"
          },
          "bins": {
            "items": {
              "$ref": "#/components/schemas/HotspotBin"
            },
            "type": "array",
            "title": "Bins"
          }
        },
        "type": "object",
        "required": [
          "grain",
          "cell_size_degrees",
          "bins"
        ],
        "title": "HotspotsResponse"
      },
...
      "IncidentPoint": {
        "properties": {
          "case_master_id": {
            "type": "integer",
            "title": "Case Master Id"
          },
          "crime_no": {
            "type": "string",
            "title": "Crime No"
          },
          "case_no": {
            "type": "string",
            "title": "Case No"
          },
          "police_station_id": {
            "type": "integer",
            "title": "Police Station Id"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "case_status_id": {
            "type": "integer",
            "title": "Case Status Id"
          },
          "crime_major_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Major Head Id"
          },
          "latitude": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Latitude"
          },
          "longitude": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Longitude"
          },
          "incident_from_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident From Date"
          },
          "crime_registered_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Registered Date"
          }
        },
        "type": "object",
        "required": [
          "case_master_id",
          "crime_no",
          "case_no",
          "police_station_id",
          "case_status_id",
          "latitude",
          "longitude"
        ],
        "title": "IncidentPoint"
      },
      "IncidentPointsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentPoint"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "IncidentPointsResponse"
      },
...
      "TrendAlert": {
        "properties": {
          "district_id": {
            "type": "integer",
            "title": "District Id"
          },
          "district_name": {
            "type": "string",
            "title": "District Name"
          },
          "crime_major_head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crime Major Head Id"
          },
          "crime_head_name": {
            "type": "string",
            "title": "Crime Head Name"
          },
          "recent_count": {
            "type": "integer",
            "title": "Recent Count"
          },
          "baseline_avg": {
            "type": "number",
            "title": "Baseline Avg"
          },
          "spike_ratio": {
            "type": "number",
            "title": "Spike Ratio"
          },
          "is_alert": {
            "type": "boolean",
            "title": "Is Alert"
          }
        },
        "type": "object",
        "required": [
          "district_id",
          "district_name",
          "crime_major_head_id",
          "crime_head_name",
          "recent_count",
          "baseline_avg",
          "spike_ratio",
          "is_alert"
        ],
        "title": "TrendAlert"
      },
      "TrendAlertsResponse": {
        "properties": {
          "recent_days": {
            "type": "integer",
            "title": "Recent Days"
          },
          "baseline_days": {
            "type": "integer",
            "title": "Baseline Days"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/TrendAlert"
            },
            "type": "array",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "recent_days",
          "baseline_days",
          "threshold",
          "alerts"
        ],
        "title": "TrendAlertsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}